models/eye_classification/fully_separated/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
dill==0.3.8
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
onnx==1.16.1
onnxruntime-gpu==1.18.0
packaging==24.0
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
timm==1.0.3
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
